requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
redis==5.2.*
qdrant-client==1.13.*
neo4j==5.27.*
langchain==0.3.*
langgraph==0.2.*
langchain-openai==0.3.*
langchain-qdrant==0.2.*
langchain-neo4j==0.2.*
langchain-google-genai==2.0.*
hdbscan==0.8.*
pydantic-settings==2.7.*
line-bot-sdk==3.5.*
apscheduler==3.10.*
scikit-learn
torch
torch_geometric
geoopt
